web-vcc/backend/requirements.txt
redis==4.4.0
types-redis==4.3.21.6
#uvloop==0.16.0
vcc-lib
multidict==6.0.4
sanic
limits
PyJWT==2.6.0
py-vapid==1.9.0
pywebpush==1.14.0
json-rpc==1.15.0